updating for jazzy
