+++ neo_rox_moveit2/config/ur10/joint_limits.yaml
@@ -1,0 +1,25 @@
+# These limits are used by MoveIt and augment/override the definitions in ur_description.
+#
+# While the robot does not inherently have any limits on joint accelerations (only on torques),
+# MoveIt needs them for time parametrization. They were chosen conservatively to work in most use
+# cases. For specific applications, higher values might lead to better execution performance.
+
+joint_limits:
+  ur10shoulder_pan_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur10shoulder_lift_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur10elbow_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur10wrist_1_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur10wrist_2_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur10wrist_3_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
