+++ neo_rox_moveit2/config/ur5/controllers.yaml
@@ -1,0 +1,38 @@
+# MoveIt uses this configuration for controller management
+moveit_controller_manager: moveit_simple_controller_manager/MoveItSimpleControllerManager
+
+trajectory_execution:
+  allowed_execution_duration_scaling: 1.2
+  allowed_goal_duration_margin: 0.5
+  allowed_start_tolerance: 0.01
+  execution_duration_monitoring: false # May lead to unexpectedly aborted goals with scaled JTC
+
+moveit_simple_controller_manager:
+  controller_names:
+    - scaled_joint_trajectory_controller
+    - joint_trajectory_controller
+
+  scaled_joint_trajectory_controller:
+    action_ns: follow_joint_trajectory
+    type: FollowJointTrajectory
+    default: true
+    joints:
+      - ur5shoulder_pan_joint
+      - ur5shoulder_lift_joint
+      - ur5elbow_joint
+      - ur5wrist_1_joint
+      - ur5wrist_2_joint
+      - ur5wrist_3_joint
+
+
+  joint_trajectory_controller:
+    action_ns: follow_joint_trajectory
+    type: FollowJointTrajectory
+    default: false
+    joints:
+      - ur5shoulder_pan_joint
+      - ur5shoulder_lift_joint
+      - ur5elbow_joint
+      - ur5wrist_1_joint
+      - ur5wrist_2_joint
+      - ur5wrist_3_joint
