+++ neo_rox_moveit2/config/ur5/joint_limits.yaml
@@ -1,0 +1,25 @@
+# These limits are used by MoveIt and augment/override the definitions in ur_description.
+#
+# While the robot does not inherently have any limits on joint accelerations (only on torques),
+# MoveIt needs them for time parametrization. They were chosen conservatively to work in most use
+# cases. For specific applications, higher values might lead to better execution performance.
+
+joint_limits:
+  ur5shoulder_pan_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur5shoulder_lift_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur5elbow_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur5wrist_1_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur5wrist_2_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
+  ur5wrist_3_joint:
+    has_acceleration_limits: true
+    max_acceleration: 5.0
