+++ neo_rox_moveit2/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>neo_ur_moveit_config</name>
+  <name>neo_rox_moveit2</name>
   <version>2.2.10</version>
   <description>
      Moveit package for mobile manipulator MMO 700
