+++ neo_rox_moveit2/rviz/view_robot.rviz
