+++ neo_rox_moveit2/srdf/rox.srdf.xacro
