--- neo_ur_moveit_config/config/kinematics.yaml
@@ -1,8 +1,0 @@
-/**:
-  ros__parameters:
-    robot_description_kinematics:
-      ur_manipulator:
-        kinematics_solver: kdl_kinematics_plugin/KDLKinematicsPlugin
-        kinematics_solver_search_resolution: 0.005
-        kinematics_solver_timeout: 0.005
-        kinematics_solver_attempts: 3
