--- neo_ur_moveit_config/config/ur10_controllers.yaml
@@ -1,29 +1,0 @@
-controller_names:
-  - scaled_joint_trajectory_controller
-  - joint_trajectory_controller
-
-
-scaled_joint_trajectory_controller:
-  action_ns: follow_joint_trajectory
-  type: FollowJointTrajectory
-  default: true
-  joints:
-    - ur10shoulder_pan_joint
-    - ur10shoulder_lift_joint
-    - ur10elbow_joint
-    - ur10wrist_1_joint
-    - ur10wrist_2_joint
-    - ur10wrist_3_joint
-
-
-joint_trajectory_controller:
-  action_ns: follow_joint_trajectory
-  type: FollowJointTrajectory
-  default: false
-  joints:
-    - ur10shoulder_pan_joint
-    - ur10shoulder_lift_joint
-    - ur10elbow_joint
-    - ur10wrist_1_joint
-    - ur10wrist_2_joint
-    - ur10wrist_3_joint
