--- neo_ur_moveit_config/config/ur5e_controllers.yaml
@@ -1,29 +1,0 @@
-controller_names:
-  - scaled_joint_trajectory_controller
-  - joint_trajectory_controller
-
-
-scaled_joint_trajectory_controller:
-  action_ns: follow_joint_trajectory
-  type: FollowJointTrajectory
-  default: true
-  joints:
-    - ur5eshoulder_pan_joint
-    - ur5eshoulder_lift_joint
-    - ur5eelbow_joint
-    - ur5ewrist_1_joint
-    - ur5ewrist_2_joint
-    - ur5ewrist_3_joint
-
-
-joint_trajectory_controller:
-  action_ns: follow_joint_trajectory
-  type: FollowJointTrajectory
-  default: false
-  joints:
-    - ur5eshoulder_pan_joint
-    - ur5eshoulder_lift_joint
-    - ur5eelbow_joint
-    - ur5ewrist_1_joint
-    - ur5ewrist_2_joint
-    - ur5ewrist_3_joint
